refactor: renamed module from data to builders CrystalGraphBuilder and HeteroGraphBuilder classes
