--- matgraphdb/pyg/data/__init__.py
@@ -1,2 +1,0 @@
-from matgraphdb.pyg.data.crystal_graph import CrystalGraphBuilder
-from matgraphdb.pyg.data.hetero_graph import HeteroGraphBuilder
