+++ matgraphdb/pyg/models/cg_target/run.py
@@ -6,9 +6,9 @@ from torch_geometric.nn import CGConv, global_mean_pool
 
 from matgraphdb import config
 from matgraphdb.core.datasets.mp_near_hull import MPNearHull
+from matgraphdb.pyg.builders import CrystalGraphBuilder
 from matgraphdb.pyg.core import BaseTrainer
 from matgraphdb.pyg.core.experiment import run_experiment
-from matgraphdb.pyg.data import CrystalGraphBuilder
 from matgraphdb.pyg.models.cg_target.model import CGConvModel
 
 print(torch.__version__)
