+++ matgraphdb/pyg/models/cgae/train.py
@@ -22,7 +22,7 @@ from torch_geometric.utils import (
 )
 
 from matgraphdb.core.datasets.mp_near_hull import MPNearHull
-from matgraphdb.pyg.data.crystal_graph import CrystalGraphBuilder
+from matgraphdb.pyg.builders.crystal_graph import CrystalGraphBuilder
 
 from .model import CGAE
