+++ matgraphdb/pyg/models/heterograph_encoder_general/train_linear.py
@@ -25,7 +25,7 @@ from sklearn.metrics import (
 from torch_geometric import nn as pyg_nn
 
 from matgraphdb.core.datasets.mp_near_hull import MPNearHull
-from matgraphdb.pyg.data import HeteroGraphBuilder
+from matgraphdb.pyg.builders import HeteroGraphBuilder
 from matgraphdb.pyg.models.heterograph_encoder_general.metrics import (
     LearningCurve,
     ROCCurve,
