+++ matgraphdb/pyg/models/hgt/train.py
@@ -15,7 +15,7 @@ from omegaconf import OmegaConf
 from torch_geometric import nn as pyg_nn
 
 from matgraphdb.core.datasets.mp_near_hull import MPNearHull
-from matgraphdb.pyg.data import HeteroGraphBuilder
+from matgraphdb.pyg.builders import HeteroGraphBuilder
 from matgraphdb.pyg.models.hgt.model import HGT
 from matgraphdb.pyg.models.hgt.trainer import (
     Trainer,
