+++ matgraphdb/pyg/models/transe/train.py
@@ -22,7 +22,7 @@ from torch_geometric import nn as pyg_nn
 from torch_geometric.nn import MetaPath2Vec, TransE
 
 from matgraphdb.core.datasets.mp_near_hull import MPNearHull
-from matgraphdb.pyg.data import HeteroGraphBuilder
+from matgraphdb.pyg.builders import HeteroGraphBuilder
 from matgraphdb.pyg.models.metapath2vec.metrics import plot_pca
 from matgraphdb.utils.colors import DEFAULT_CMAP, DEFAULT_COLORS
 from matgraphdb.utils.config import config
